requirements.txt
google-generativeai>=0.3.2
requests>=2.31.0
pathlib>=1.0.1
typing-extensions>=4.9.0 